extracts useLinks hook
